backend/requirements/dev.txt
aioredis==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
bcrypt==4.1.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
email-validator==2.0.0.post2
fastapi==0.105.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.2
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hiredis==2.2.3
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
httpx-oauth==0.13.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
libclang==16.0.6
makefun==1.15.2
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pendulum==2.1.2
Pillow==10.1.0
prometheus-client==0.19.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.39.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
starlette-exporter==0.17.1
termcolor==2.4.0
text-unidecode==1.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
uvicorn
fastapi-cache2==0.2.1